pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cnc-machine-core"
version = "0.1.0"
description = "Core CNC machine control module"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml",
  "pyserial",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["cnc_machine_core"]

[tool.setuptools.package-data]
cnc_machine_core = ["deck/*.json", "labware/*.json"]
